Revert "forward declare tbb to resolve link error"
This reverts commit 5a234d6ae247156cb490c3c60f992765016f0b0b.
